refactor: 커스텀 예외처리를 표준 예외로 변경
InputException --> IllegalArgument Exception
